+++ llama-index-networks/llama_index/networks/__init__.py
@@ -1,11 +1,11 @@
-from llama_index.networks.contributor import (
+from llama_index.networks.contributor.query_engine import (
     ContributorClient,
     ContributorClientSettings,
     ContributorService,
     ContributorServiceSettings,
 )
 
-from llama_index.networks.query_engine import NetworkQueryEngine
+from llama_index.networks.network.query_engine import NetworkQueryEngine
 
 __all__ = [
     "ContributorClient",
