--- llama-index-networks/llama_index/networks/query_engine/BUILD
@@ -1,1 +1,0 @@
-python_sources()
