--- llama-index-networks/llama_index/networks/query_engine/__init__.py
@@ -1,4 +1,0 @@
-from llama_index.networks.query_engine.network_query_engine import NetworkQueryEngine
-
-
-__all__ = ["NetworkQueryEngine"]
