+++ llama-index-networks/tests/contributor/query_engine/test_service.py
@@ -1,5 +1,5 @@
 from fastapi.testclient import TestClient
-from llama_index.networks import (
+from llama_index.networks.contributor.query_engine import (
     ContributorService,
     ContributorServiceSettings,
 )
